raw/eks-control-plane-custom-resource-function/requirements.txt
crhelper
sockets
testresources
pyopenssl
certifi
